Change batch processing file hierarchy
